requirements.txt
Flask
gunicorn
google-cloud-storage
google-cloud-aiplatform
Pillow
fpdf2